refactor(excel-checker): moves math spreadsheet services in math package
